Update pytest dependencies and configuration for better compatibility.
We now install the right versions of `pytest` and `pytest-env` for the
right versions of Python.

Pytest is now configured to require the necessary plugins before it
runs, and to test explicitly the files, functions, and classes we want
it to test.

Testing Done:
Successfully installed the right packages in a series of virtualenvs.

All unit tests pass for all supported versions of Django and Python.

Verified both with tox.

Reviewed at https://reviews.reviewboard.org/r/11995/