Use getfullargspec on Python 3 and getargspec on Python 2.
We need to check the arguments of the function passed into a
`SQLMutation`, to determine if we need to be in a legacy mode when
calling that function. We were using `getargspec()` for this, but that
emits deprecation warnings on Python 3.

We now call `getfullargspec()` on Python 3 instead, avoiding these
warnings.

Testing Done:
All unit tests pass for all supported configurations.

Reviewed at https://reviews.reviewboard.org/r/11996/