Add a module for placeholders for hinted evolutions.
When generating hinted evolutions, sometimes we're unable to compute a
value and instead place a marker that users must fill in. This ensures a
syntax error is hit, rather than executing a bad evolution.

There's currently only one placeholder type, `NullFieldInitialCallback`.
This change gives it a parent class, `BasePlaceholder`, and moves back
into a new `placeholders.py`. This reduces dependence in the codebase on
the specific class, and avoids an upcoming circular dependency caused by
having it in `diff.py`.

Testing Done:
Unit tests pass on all supported versions of Python and Django.

Tested this with an upcoming change that checks against `BasePlaceholder`.

Reviewed at https://reviews.reviewboard.org/r/11982/