Allow database backends to specify their default tablespace.
Django 3.2+'s `Index` support allows a tablespace to be set for
databases that support it. In preparation, this change introduces a
`default_tablespace` argument on the database evolver backend, and a
testing function for querying it, which can be used during test runs.

Testing Done:
Unit tests pass for all supported versions of Django and for all
databases. This isn't used yet, though.

Reviewed at https://reviews.reviewboard.org/r/12042/