Add testing decorators for requiring parameters or attributes.
This introduces `@requires_attr` and `@requires_param`, for requiring an
attribute on a class or a parameter to a function, respectively.

These can be used by unit tests to skip tests if some object or function
in the running version of Django doesn't include a particular feature.

These decorators will be used in upcoming changes.

Testing Done:
Tested with upcoming changes that use these decorators.

Reviewed at https://reviews.reviewboard.org/r/12044/