Fix running unit tests through the old test runner.
The old test runner (currently used in CI) had a regression when moving
to pytest. It failed to convert the first argument (the database name to
run everything against) to the new `pytest --db=` argument.

This change adds this back to the test runner. If an argument is passed,
and isn't `--`, it's expected to be the database name as before. All
other arguments are then passed as normal.

Testing Done:
Ran the old test runner with/without a database name, with/without `--`,
and with/without specific tests to run.

Reviewed at https://reviews.reviewboard.org/r/12129/