Add classes for warning on deprecations.
This introduces a new `django_evolution.deprecation` module, which
contains utility classes that warn about usage of deprecated
functionality. This includes the following stable classes:

* `BaseRemovedInDjangoEvolutionWarning`: Base class for all deprecation
  warnings.

* `RemovedInNextDjangoEvolutionWarning`: Indicates removal in the next
  major version of Django Evolution.

And version-specific classes, which are current as of Django Evolution
2.2 but will be added to and removed as needed:

* `RemovedInDjangoEvolution30Warning`: Deprecation warning for features
  scheduled to be removed in Django Evolution 3.0.

* `RemovedInDjangoEvolution40Warning`: Deprecation warning for features
  scheduled to be removed in Django Evolution 4.0.

`RemovedInNextDjangoEvolutionWarning` currently maps to
`RemovedInDjangoEvolution30Warning.

As of this change, there are no features flagged for deprecation, but
an upcoming change will begin to introduce deprecations.

Testing Done:
Made use of this in an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/12217/