Allow get_evolution_dependencies to take a list of custom evolutions.
`get_evolution_dependencies()` can now be passed a list of custom
evolutions, which will be searched if a module for the evolution cannot
be found. This is in the same format that `EvolveAppTask` takes.
Dependencies can be specified within each custom evolution or through
listed mutations (such as `MoveToDjangoMigrations`).

This will soon be used by the evolver, primarily for the purpose of unit
testing.

Testing Done:
Unit tests pass.

Tested this along with pending changes to the evolver.

Reviewed at https://reviews.reviewboard.org/r/12267/