Split django_evolution.mutators into multiple sub-modules.
In preparation for some upcoming work on mutators, this change splits up
the somewhat large `django_evolution.mutators` module into per-class
modules. This will help keep things organized as new mutator classes are
added (which will be needed for the new `MoveToDjangoEvolution`
mutation).

The actual code in these modules is untouched, with the exception of
the addition of module-level strings, a `Version Changed` in each class,
and a new docstring for `SQLMutator`. Docs otherwise are not up to
current standards.

Testing Done:
All unit tests pass.

Built the docs and checked for errors.

Reviewed at https://reviews.reviewboard.org/r/12269/