Differentiate database installs from upgrades in success messages.
When running the evolver, we've historically shown the message "The
database upgrade was successful!" This made sense back before the entire
database setup and upgrade process was controlled by Django Evolution.
Now, since it handles population of the initial database, this message
isn't enough.

We now track whether this appears to be a brand-new install vs. an
upgrade (based on the presence of the `Version` and `Evolution` models
in the database), and use that to determine which message to display.

Testing Done:
Tested installing a new database and upgrading an old one. Saw the
correct message in both circumstances.

Reviewed at https://reviews.reviewboard.org/r/12313/