Add release notes for Django Evolution 2.2.
This is a pretty big release under the hood, mostly centered around code
cleanup, Django 3.2/4.0 compatibility, and with a lot of fixes to index
management (mostly to faciliate the `Index` class).

As the vast majority of the work is internal, and some was for the
il-fated Custom Migrations support, those have been left out of the
release notes.

Testing Done:
Built the docs. Checked for spelling, build, and link errors.

Reviewed at https://reviews.reviewboard.org/r/12389/