Fix configuration/environmental issues for unit tests.
We now disable pytest-django, which interferes with our test setup when
installed (since pytest auto-enables plugins). This fixes test failures
when not in an isolated environment.

The postgres11.8 environments had a badly-formatted name, causing it to
include a `{` in its name. This now uses `_` instead of `.` in the name.

Testing Done:
Unit tests pass.

Verified the name of the environments were correct.

Reviewed at https://reviews.reviewboard.org/r/13092/