Fix the formatting of SQL test data for Postgres on Django 1.6.
Django 1.6's generated SQL output for Postgres CREATE TABLE operations
differs greatly from that of future versions of Django, and for other
types of operations. Instead of a single string for a whole CREATE
TABLE, we get one line of each in the generated SQL array.

Because of this difference, my update to add `pk_field_def` failed,
since I added it to the end of a CREATE TABLE block and not to the line
that needed it.

This is a trivial change, updating just a few statements.