Always consider new migration-aware apps when evolving.
The code for building the migration graph wasn't considering
migration-aware apps that didn't already exist in the database. This
could prevent some migrations from applying or being considered for
custom evolutions.

The cause was a small logic error where we tried to avoid working with
the signature if it didn't exist, but we checked this too early and
failed to put the app in the list of migration-aware apps. The logic has
moved to the appropriate conditional.

Testing Done:
All unit tests passed.

Verified that this fixed an issue with an newly-added app not being
considered for migration.

Reviewed at https://reviews.reviewboard.org/r/14214/