Update test matrix.
This change updates the django-evolution test matrix for the modern
world. This includes updates to both the tox configuration as well as
the DB settings and docker-compose setup used for database-specific
tests.

- Python support now ranges from 3.10 to 3.14
- Django support now ranges from 4.0 to 6.0
- MySQL now includes 5.7, 8, and 9. 5.6 support was dropped in Django
  3.2, so we no longer need to test it.
- Mariadb now has separate containers for 10, 11, and 12.
- Postgres now ranges from 11 through 18.

I've added some additional comments to the tox config making note of the
database deprecation schedule from Django's releases so we can get rid
of old versions as we do future updates.

I briefly flirted with trying to port the tox config to the newer
toml-based format, but gave up for now. That may come again in a later
change.

Testing Done:
Ran full test suite (from the top of my commit stack, including a
variety of updates and fixes for new Django versions).

Reviewed at https://reviews.reviewboard.org/r/14738/