Move django-evolution to pyproject.toml and set minimum dependencies.
This change moves django-evolution over to the new pyproject.toml setup
for Python packages. This is a relatively straightforward update, which
can use the regular setuptools build backend.

As part of this next major release, we'll be setting a floor for
dependencies and cutting off a bunch of old compatibility. Given the
release timeline, I think Python 3.10 is the correct minimum version to
support. While we'll also be using Django 5.2 as of then, for now I've
chosen to set the minimum Django version as 4.0, because I feel like
there's a lot of usefulness for developers in having some overlap in
Django support between django-evolution 2.4.2 and 3.0.

This also updates our LICENSE file a tiny bit to update copyright dates,
as well as clarify the names in the third clause of the license.

Testing Done:
Did an editable install and built a wheel.

Reviewed at https://reviews.reviewboard.org/r/14737/