Fix test output to show failures in the summary.
The default command-line options configured for pytest were showing
explicit skips, but did not also include the default failure/error
states, so failed test cases would not be listed in the final summary.

This change fixes that up so that we include the default `fE` plus the
desired `s` flags.

Testing Done:
Ran unit tests and saw the list of failures in the final summary.

Reviewed at https://reviews.reviewboard.org/r/14739/